backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "fastapi>=0.121.2",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.2.12",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.4",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0",
    "uvicorn[standard]>=0.38.0",
]
